# frontend/src/types/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "werewolf-agent-sim server",
    "version": "0.1.0"
  },
  "paths": {
...
    "/rooms": {
      "get": {
        "tags": [
          "rooms"
        ],
        "summary": "Browse waiting rooms (public)",
        "description": "Return non-stale waiting rooms newest first without expiring direct URLs.",
        "operationId": "list_rooms_rooms_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RoomSummary"
                  },
                  "type": "array",
                  "title": "Response List Rooms Rooms Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "rooms"
        ],
        "summary": "Create a multiplayer waiting room (humans join via /join)",
        "description": "Create a lobby whose identifier remains stable when the game starts.\n\nHumans enter only through ``/join`` and roles remain random. The model is resolved\nand the house consulted here so the host learns early; the cap is checked again at\nstart, which is when the game actually costs anything.",
        "operationId": "create_room_rooms_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewRoom"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomCreated"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/games/{game_id}/lock": {
      "post": {
        "tags": [
          "rooms"
        ],
        "summary": "Lock or unlock a waiting room (host only)",
        "operationId": "lock_room_games__game_id__lock_post",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Game Id"
            }
          },
          {
            "name": "host_key",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Host Key"
            }
          },
          {
            "name": "locked",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Locked"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/games/{game_id}/join": {
      "post": {
        "tags": [
          "rooms"
        ],
        "summary": "Claim a human seat in a waiting room",
        "operationId": "join_game_games__game_id__join_post",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Game Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinGame"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeatJoined"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/games/{game_id}/rejoin": {
      "post": {
        "tags": [
          "rooms"
        ],
        "summary": "Restore a lost seat cookie from the token's body copy",
        "operationId": "rejoin_game_games__game_id__rejoin_post",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Game Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejoinGame"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeatJoined"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/games/{game_id}/start": {
      "post": {
        "tags": [
          "rooms"
        ],
        "summary": "Start the waiting room's game (host only)",
        "description": "Replace the lobby with a running session under the same ID (registry.start).",
        "operationId": "start_game_games__game_id__start_post",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Game Id"
            }
          },
          {
            "name": "host_key",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Host Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameCreated"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/replays": {
      "get": {
        "tags": [
          "replays"
        ],
        "summary": "Browse finished games (newest first)",
        "description": "One page of finished games. The body stays a plain list; the total number of\nfinished games, for a footer like \"212 games\", rides in the X-Total-Count header.",
        "operationId": "list_replays_replays_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReplayBase"
                  },
                  "title": "Response List Replays Replays Get"
                }
              }
            },
            "headers": {
              "X-Total-Count": {
                "description": "How many finished games there are in all, across every page.",
                "schema": {
                  "type": "integer"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/replays/{game_id}": {
      "get": {
        "tags": [
          "replays"
        ],
        "summary": "One finished game's full event log",
        "operationId": "get_replay_replays__game_id__get",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Game Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplayGame"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GameCreated": {
        "properties": {
          "game_id": {
            "type": "string",
            "title": "Game Id"
          },
          "seat_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seat Token"
          }
        },
        "type": "object",
        "required": [
          "game_id"
        ],
        "title": "GameCreated",
        "description": "POST /games and POST /games/{id}/start response."
      },
...
      "GameStarted": {
        "properties": {
          "seq": {
            "type": "integer",
            "title": "Seq"
          },
          "day": {
            "type": "integer",
            "title": "Day"
          },
          "type": {
            "type": "string",
            "const": "game_started",
            "title": "Type",
            "default": "game_started"
          },
          "seats": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Seats"
          },
          "cast_role_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Cast Role Counts"
          }
        },
        "type": "object",
        "required": [
          "seq",
          "day",
          "seats",
          "cast_role_counts"
        ],
        "title": "GameStarted"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JoinGame": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "default": "human"
          }
        },
        "type": "object",
        "title": "JoinGame",
        "description": "POST /games/{id}/join body. No role field on purpose: rooms deal random\nseats (role choice is solo-only, on the instant-start path)."
      },
...
      "NewRoom": {
        "properties": {
          "api_key": {
            "type": "string",
            "title": "Api Key",
            "default": ""
          },
          "model": {
            "type": "string",
            "title": "Model",
            "default": ""
          },
          "memory": {
            "type": "boolean",
            "title": "Memory",
            "default": false
          },
          "name": {
            "type": "string",
            "maxLength": 40,
            "title": "Name",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "NewRoom",
        "description": "POST /rooms body: the multiplayer door. No human/role fields ON PURPOSE \u2014\na room seats humans only via POST /join and always deals random roles, so the\nconcept is unrepresentable here instead of guarded; extra=\"forbid\" turns a\nclient sending instant-start fields into a clean 422."
      },
...
      "RejoinGame": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          }
        },
        "type": "object",
        "required": [
          "token"
        ],
        "title": "RejoinGame",
        "description": "POST /games/{id}/rejoin body: re-prove seat ownership after cookie loss\n(new device, cleared browsing data) with the stashed body-copy token."
      },
      "ReplayBase": {
        "properties": {
          "game_id": {
            "type": "string",
            "title": "Game Id"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "winner": {
            "type": "string",
            "enum": [
              "villagers",
              "wolves",
              "serial_killer"
            ],
            "title": "Winner"
          },
          "days": {
            "type": "integer",
            "title": "Days"
          },
          "ended_phase": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "day",
                  "voting",
                  "night"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended Phase"
          },
          "n_events": {
            "type": "integer",
            "title": "N Events"
          },
          "n_humans": {
            "type": "integer",
            "title": "N Humans"
          },
          "cast_role_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Cast Role Counts"
          },
          "model": {
            "type": "string",
            "title": "Model",
            "default": ""
          },
          "memory": {
            "type": "boolean",
            "title": "Memory",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "game_id",
          "winner",
          "days",
          "n_events",
          "n_humans",
          "cast_role_counts"
        ],
        "title": "ReplayBase",
        "description": "One row returned by ``GET /replays``."
      },
      "ReplayGame": {
        "properties": {
          "game_id": {
            "type": "string",
            "title": "Game Id"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "winner": {
            "type": "string",
            "enum": [
              "villagers",
              "wolves",
              "serial_killer"
            ],
            "title": "Winner"
          },
          "days": {
            "type": "integer",
            "title": "Days"
          },
          "ended_phase": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "day",
                  "voting",
                  "night"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended Phase"
          },
          "n_events": {
            "type": "integer",
            "title": "N Events"
          },
          "n_humans": {
            "type": "integer",
            "title": "N Humans"
          },
          "cast_role_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Cast Role Counts"
          },
          "model": {
            "type": "string",
            "title": "Model",
            "default": ""
          },
          "memory": {
            "type": "boolean",
            "title": "Memory",
            "default": false
          },
          "events": {
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/GameStarted"
                },
                {
                  "$ref": "#/components/schemas/RoleAssigned"
                },
                {
                  "$ref": "#/components/schemas/RolesAssigned"
                },
                {
                  "$ref": "#/components/schemas/PhaseChange"
                },
                {
                  "$ref": "#/components/schemas/GameOver"
                },
                {
                  "$ref": "#/components/schemas/TurnStarted"
                },
                {
                  "$ref": "#/components/schemas/Speech"
                },
                {
                  "$ref": "#/components/schemas/PassMarker"
                },
                {
                  "$ref": "#/components/schemas/FiringReasonAnnotation"
                },
                {
                  "$ref": "#/components/schemas/AddressedTargetsAnnotation"
                },
                {
                  "$ref": "#/components/schemas/StrategyUpdate"
                },
                {
                  "$ref": "#/components/schemas/MemoryConsulted"
                },
                {
                  "$ref": "#/components/schemas/MemoryExtracted"
                },
                {
                  "$ref": "#/components/schemas/PlayerReads"
                },
                {
                  "$ref": "#/components/schemas/InputRequest"
                },
                {
                  "$ref": "#/components/schemas/DaySummary"
                },
                {
                  "$ref": "#/components/schemas/DaySummaryStructured"
                },
                {
                  "$ref": "#/components/schemas/VoteCast"
                },
                {
                  "$ref": "#/components/schemas/GmMessage"
                },
                {
                  "$ref": "#/components/schemas/LynchResult"
                },
                {
                  "$ref": "#/components/schemas/RosterUpdate"
                },
                {
                  "$ref": "#/components/schemas/PackRosterUpdate"
                },
                {
                  "$ref": "#/components/schemas/NightAction"
                },
                {
                  "$ref": "#/components/schemas/WolfMessage"
                },
                {
                  "$ref": "#/components/schemas/WolfVote"
                },
                {
                  "$ref": "#/components/schemas/WolfKillDecided"
                },
                {
                  "$ref": "#/components/schemas/NightResult"
                },
                {
                  "$ref": "#/components/schemas/InvestigationResult"
                },
                {
                  "$ref": "#/components/schemas/VigilanteConfirmation"
                },
                {
                  "$ref": "#/components/schemas/BulletsRemaining"
                }
              ],
              "discriminator": {
                "propertyName": "type",
                "mapping": {
                  "addressed_targets": "#/components/schemas/AddressedTargetsAnnotation",
                  "bullets_remaining": "#/components/schemas/BulletsRemaining",
                  "day_summary": "#/components/schemas/DaySummary",
                  "day_summary_structured": "#/components/schemas/DaySummaryStructured",
                  "firing_reason": "#/components/schemas/FiringReasonAnnotation",
                  "game_over": "#/components/schemas/GameOver",
                  "game_started": "#/components/schemas/GameStarted",
                  "gm_message": "#/components/schemas/GmMessage",
                  "input_request": "#/components/schemas/InputRequest",
                  "investigation_result": "#/components/schemas/InvestigationResult",
                  "lynch_result": "#/components/schemas/LynchResult",
                  "memory_consulted": "#/components/schemas/MemoryConsulted",
                  "memory_extracted": "#/components/schemas/MemoryExtracted",
                  "night_action": "#/components/schemas/NightAction",
                  "night_result": "#/components/schemas/NightResult",
                  "pack_roster_update": "#/components/schemas/PackRosterUpdate",
                  "pass_marker": "#/components/schemas/PassMarker",
                  "phase_change": "#/components/schemas/PhaseChange",
                  "player_reads": "#/components/schemas/PlayerReads",
                  "role_assigned": "#/components/schemas/RoleAssigned",
                  "roles_assigned": "#/components/schemas/RolesAssigned",
                  "roster_update": "#/components/schemas/RosterUpdate",
                  "speech": "#/components/schemas/Speech",
                  "strategy_update": "#/components/schemas/StrategyUpdate",
                  "turn_started": "#/components/schemas/TurnStarted",
                  "vigilante_confirmation": "#/components/schemas/VigilanteConfirmation",
                  "vote_cast": "#/components/schemas/VoteCast",
                  "wolf_kill_decided": "#/components/schemas/WolfKillDecided",
                  "wolf_message": "#/components/schemas/WolfMessage",
                  "wolf_vote": "#/components/schemas/WolfVote"
                }
              }
            },
            "type": "array",
            "title": "Events"
          }
        },
        "type": "object",
        "required": [
          "game_id",
          "winner",
          "days",
          "n_events",
          "n_humans",
          "cast_role_counts",
          "events"
        ],
        "title": "ReplayGame",
        "description": "One complete replay with every stored event revalidated for the wire."
      },
...
      "RolesAssigned": {
        "properties": {
          "seq": {
            "type": "integer",
            "title": "Seq"
          },
          "day": {
            "type": "integer",
            "title": "Day"
          },
          "type": {
            "type": "string",
            "const": "roles_assigned",
            "title": "Type",
            "default": "roles_assigned"
          },
          "roles": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Roles"
          }
        },
        "type": "object",
        "required": [
          "seq",
          "day",
          "roles"
        ],
        "title": "RolesAssigned",
        "description": "O-tier full deal at game start: nobody receives it live; exists so the game_over\nbacklog replay has survivor roles from minute zero."
      },
      "RoomCreated": {
        "properties": {
          "game_id": {
            "type": "string",
            "title": "Game Id"
          },
          "host_key": {
            "type": "string",
            "title": "Host Key"
          }
        },
        "type": "object",
        "required": [
          "game_id",
          "host_key"
        ],
        "title": "RoomCreated",
        "description": "POST /rooms response."
      },
      "RoomSummary": {
        "properties": {
          "game_id": {
            "type": "string",
            "title": "Game Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "players": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Players"
          },
          "host": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Host"
          },
          "max_seats": {
            "type": "integer",
            "title": "Max Seats"
          },
          "locked": {
            "type": "boolean",
            "title": "Locked"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "game_id",
          "name",
          "players",
          "max_seats",
          "locked",
          "created_at"
        ],
        "title": "RoomSummary",
        "description": "One row of GET /rooms: the public face of a waiting room. Also the POST /lock\nresponse (the host's fresh view after flipping the flag). Never any secret \u2014\nno host_key, no tokens."
      },
...
      "SeatJoined": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          }
        },
        "type": "object",
        "required": [
          "token"
        ],
        "title": "SeatJoined",
        "description": "POST /games/{id}/join and POST /games/{id}/rejoin response."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}